에러처리 완료

+++ src/main/java/com/mju/Board/application/BoardServiceImpl.java
@@ -2,9 +2,8 @@ package com.mju.Board.application;
 
 import com.mju.Board.domain.model.Exception.ExceptionList;
 import com.mju.Board.domain.model.FAQBoard;
-import com.mju.Board.domain.model.Result.CommonResult;
 import com.mju.Board.domain.repository.FAQBoardRepository;
-import com.mju.Board.domain.service.ResponseService;
+import com.mju.Board.domain.model.Exception.FaqBoardNotFindException;
 import com.mju.Board.presentation.dto.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -12,8 +11,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-
-import static com.mju.Board.domain.model.Exception.ExceptionList.NOT_EXISTENT_FAQBOARD;
 
 
 @Service
@@ -24,6 +21,7 @@ public class BoardServiceImpl implements BoardService{
     @Override
     @Transactional
     public void registerFaqGeneral(FAQRegisterDto faqRegisterDto) {
+        try {
         FAQBoard originalFaqBoard = new FAQBoard(faqRegisterDto.getFaqTitle(), faqRegisterDto.getFaqContent(),faqRegisterDto.isChecked());
         FAQBoard newFaqBoard = FAQBoard.builder()
                 .faqIndex(originalFaqBoard.getFaqIndex())
@@ -36,23 +34,30 @@ public class BoardServiceImpl implements BoardService{
                 .isChecked(originalFaqBoard.isChecked())
                 .build();
         faqBoardRepository.save(newFaqBoard);
+        }catch (FaqBoardNotFindException e){
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_REGISTER_GENERAL_MEMBER);
+        }
     }
 
     @Override
     @Transactional
     public void registerFaqAdu(FAQRegisterDto faqRegisterDto) {
-        FAQBoard originalFaqBoard = new FAQBoard(faqRegisterDto.getFaqTitle(), faqRegisterDto.getFaqContent(), faqRegisterDto.isChecked());
-        FAQBoard newFaqBoard = FAQBoard.builder()
-                .faqIndex(originalFaqBoard.getFaqIndex())
-                .faqTitle(originalFaqBoard.getFaqTitle())
-                .faqContent(originalFaqBoard.getFaqContent())
-                .createdAt(originalFaqBoard.getCreatedAt())
-                .updatedAt(originalFaqBoard.getUpdatedAt())
-                .type(FAQBoard.FAQType.EDUCATION)
-                .count(originalFaqBoard.getCount())
-                .isChecked(originalFaqBoard.isChecked())
-                .build();
-        faqBoardRepository.save(newFaqBoard);
+        try {
+            FAQBoard originalFaqBoard = new FAQBoard(faqRegisterDto.getFaqTitle(), faqRegisterDto.getFaqContent(), faqRegisterDto.isChecked());
+            FAQBoard newFaqBoard = FAQBoard.builder()
+                    .faqIndex(originalFaqBoard.getFaqIndex())
+                    .faqTitle(originalFaqBoard.getFaqTitle())
+                    .faqContent(originalFaqBoard.getFaqContent())
+                    .createdAt(originalFaqBoard.getCreatedAt())
+                    .updatedAt(originalFaqBoard.getUpdatedAt())
+                    .type(FAQBoard.FAQType.EDUCATION)
+                    .count(originalFaqBoard.getCount())
+                    .isChecked(originalFaqBoard.isChecked())
+                    .build();
+            faqBoardRepository.save(newFaqBoard);
+        }catch (FaqBoardNotFindException e){
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_REGISTER_EDUCATION);
+        }
     }
 
 
@@ -62,14 +67,22 @@ public class BoardServiceImpl implements BoardService{
     public List<FAQBoard> getGeneralFAQBoardList() {
 //        CommonResult commonResult = responseService.getListResult(faqBoardRepository.findByType(FAQBoard.FAQType.GENERAL_MEMBER));
         List<FAQBoard> generalFAQBoardList = faqBoardRepository.findByType(FAQBoard.FAQType.GENERAL_MEMBER);
-        return generalFAQBoardList;
+        if (!generalFAQBoardList.isEmpty()) {
+            return generalFAQBoardList;
+        }else {
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_NOT_FINDTYPE_GENERAL_MEMBER);
+        }
     }
 
     @Override
     @Transactional
     public List<FAQBoard> getAduFAQBoardList() {
         List<FAQBoard> aduFAQBoardList = faqBoardRepository.findByType(FAQBoard.FAQType.EDUCATION);
+        if (!aduFAQBoardList.isEmpty()) {
         return aduFAQBoardList;
+        }else{
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_NOT_FINDTYPE_EDUCATION);
+        }
     }
 
     @Override
@@ -77,18 +90,24 @@ public class BoardServiceImpl implements BoardService{
     public void updateFaq(Long faqIndex, FAQUpdateDto faqUpdateDto) {
         Optional<FAQBoard> optionalFaqBoard = faqBoardRepository.findById(faqIndex);
 //        FAQBoard faqBoard = optionalFaqBoard.orElseThrow(() -> new Exception("FAQBoard not found"));
+        if (optionalFaqBoard.isPresent()) {
         FAQBoard faqBoard = optionalFaqBoard.get();
         faqBoard.faqUpdate(faqUpdateDto.getFaqTitle(), faqUpdateDto.getFaqContent(), faqUpdateDto.isChecked());
         faqBoardRepository.save(faqBoard);
+        }else{
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_NOT_EXISTENT_UPDATE);
+        }
     }
 
     @Override
     @Transactional
     public void deleteFaq(Long faqIndex){
-//        Optional<FAQBoard> optionalFaqBoard = faqBoardRepository.findById(faqIndex);
-//        if (optionalFaqBoard.isPresent()) { // 해당 index의 FAQBoard가 존재하는 경우
+        Optional<FAQBoard> optionalFaqBoard = faqBoardRepository.findById(faqIndex);
+        if (optionalFaqBoard.isPresent()) { // 해당 index의 FAQBoard가 존재하는 경우
         faqBoardRepository.deleteById(faqIndex);
-//        } else { // 해당 index의 FAQBoard가 존재하지 않는 경우
+        } else { // 해당 index의 FAQBoard가 존재하지 않는 경우
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_NOT_EXISTENT_DELETE);
+        }
     }
     @Override
     @Transactional
@@ -98,6 +117,8 @@ public class BoardServiceImpl implements BoardService{
             FAQBoard faqBoard = optionalFaqBoard.get();
             faqBoard.incrementCount();
             faqBoardRepository.save(faqBoard);
+        }else {
+            throw new FaqBoardNotFindException(ExceptionList.FAQBOARD_NOT_EXISTENT_COUNT);
         }
     }
